authenticate memories routes

+++ server/src/routes/memories.ts
@@ -3,8 +3,16 @@ import { z } from "zod"
 import { prisma } from "../lib/prisma"
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get("/memories", async () => {
+  app.addHook("preHandler", async (request) => {
+    await request.jwtVerify()
+  })
+
+  app.get("/memories", async (request) => {
+    const { sub: userId } = request.user
     const memories = await prisma.memory.findMany({
+      where: {
+        userId,
+      },
       orderBy: {
         createdAt: "asc",
       },
@@ -18,7 +26,8 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
   })
 
-  app.get("/memories/:id", async (request) => {
+  app.get("/memories/:id", async (request, reply) => {
+    const { sub: userId } = request.user
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -28,10 +37,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id,
       },
     })
+    if (!memory.isPublic && memory.userId !== userId) {
+      return reply.status(401).send()
+    }
     return memory
   })
 
   app.post("/memories", async (request) => {
+    const { sub: userId } = request.user
     const bodySchema = z.object({
       content: z.string(),
       coverUrl: z.string(),
@@ -43,13 +56,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: "da7ed124-3990-4e6f-be98-56938eb26672",
+        userId,
       },
     })
     return memory
   })
 
-  app.put("/memories/:id", async (request) => {
+  app.put("/memories/:id", async (request, reply) => {
+    const { sub: userId } = request.user
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -60,7 +74,15 @@ export async function memoriesRoutes(app: FastifyInstance) {
       isPublic: z.coerce.boolean().default(false),
     })
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+    if (memory.userId !== userId) {
+      return reply.status(401).send()
+    }
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -73,11 +95,20 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory
   })
 
-  app.delete("/memories/:id", async (request) => {
+  app.delete("/memories/:id", async (request, reply) => {
+    const { sub: userId } = request.user
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
     const { id } = paramsSchema.parse(request.params)
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+    if (memory.userId !== userId) {
+      return reply.status(401).send()
+    }
     await prisma.memory.delete({
       where: {
         id,
